osd/OSDMap: lookup_pg_pool_name returns optional
Since it's either there or not and the only error is not-found, simply
return std::optional. It composes better and we don't have to
use/check for disjoint values.

Also some places would assert on any error OTHER THAN -ENOENT so it's
not like we were going to expand the possible failures.

Signed-off-by: Adam C. Emerson <aemerson@redhat.com>