mon/MonClient: send logs to mon on separate schedule than pings
We want to ping every 10s, but if we have log messages queued up, we
should send those promptly.  Adjust the tick interval to be the min of
these two intervals, and wrap each chunk of work with a check for
whether
enough time has passed.

This makes 'ceph -w' (or 'ceph -W $channel') *way* more responsive and
useful.

Signed-off-by: Kefu Chai <kchai@redhat.com>