osd/osd_types: reset head etc. if the whole pg-log list is divergent
When merging log, it is possible that we have all local log
entries falling behind with the incoming log list (usually
the authoritative history), as a result of which we may have
to throw the whole local list out and reset head to:
	MAX(incoming-log's tail, my-log's tail)

Since by definition a log tail is the cursor pointing to the
log version that is prior to oldest, it is also possible the
oldest incoming log entry could have a version greater than
the new local head. For example::

	incoming-log: (19'4620,34'6006], oldest(33'4621)
        local-log:(19'5300,19'5379], oldest(19'5301)
	local-log-head(when rewinding is done) = MAX(19'4620, 19'5300) = 19'5300

As we start to merge incoming log entries into our local empty
list, OSDs will crash if they see any log entry to add has a newer
version than current log head (e.g., because PG-logs are always
strictly ordered)::

	first-log-entry-to-merge: 33'4621
	local-log-head: 19'5300

Fix by resetting log head etc. to a sane version if the whole
thing goes divergent, which are also the default values by
definition that an empty log list should have.

Fixes: https://tracker.ceph.com/issues/44532
Signed-off-by: xie xingguo <xie.xingguo@zte.com.cn>