mgr/telemetry: fix enabling and disabling telemetry
Due to syntax differences in getting / setting config options, telemetry
could not be enabled and disabled, and threw an error when:
'telemetry on sharing-1-0'

TypeError: ceph_config_set() argument 2 must be string or None, not bool

This also fixes an issue where revision was not reset when disabling the
module.

This was not be backported from master due to major differences in
config syntax (attribute-based in master vs. map-based in Luminous).

Fixes: https://tracker.ceph.com/issues/44563
Signed-off-by: Yaarit Hatuka <yaarit@redhat.com>