crimson/net: allow mark_down() inside ms_handle_reset()
Although it is not necessary to mark_down the connection in its
ms_handle_reset() event, but it can be more convenient to allow it.
And Heartbeat already encounters this assertion failure.

So move the assertion to close_clean() which will help identify problems
if we happen to make ms_handle_reset() wait for messenger shutdown.

Signed-off-by: Yingxin Cheng <yingxin.cheng@intel.com>