rgw/cloud-sync: Handle BucketAlreadyExists error
Cloud-sync module tries to create parent bucket for every
object it syncs to the remote s3 cloud server. If the bucket
alredy exists, it expects "BucketAlreadyOwnedByYou" to be
returned by the server (which seem to be the right response as per
https://docs.aws.amazon.com/AmazonS3/latest/API/ErrorResponses.html)

However there are some s3 server implementation(eg., Noobaa)
which return "BucketAlreadyExists" error instead. To be able to
handle such servers, defined a new tier-config option
"allow_bucketalreadyexists" to allow "BucketAlreadyExists" error if
needed as acceptable error code in cloud-sync module.

Signed-off-by: Soumya Koduri <skoduri@redhat.com>