NVMEDevice: set the queue_size of the ctrlr opts.
Purpose: Allocate more resoures for the requests in the io qpair,
and the real value will be determined by the inner logic in SPDK NVMe Driver.
But if we do not set it, the default size is 512, and the value is too low.
If there are many big I/Os with high QD submitted, with this setting,
it could improve the performance.

Signed-off-by: Ziye Yang <ziye.yang@intel.com>