do_cmake.sh: fix application of -DWITH_RADOSGW_KAFKA_ENDPOINT=OFF
50ba25b2aa4db34319a7b95e789065540347f895 broke do_cmake.sh in nautilus
and the backport of 46b3d671b4408a766aaed379274242ad4b38afb3 (which
was included in the same PR) did not fix it enough.

This commit cannot be backported from master. It is needed to fix
a partially broken backport.

After this commit is applied, the relevant part of do_cmake.sh in
nautilus will look exactly like it does now in master.

Fixes: 50ba25b2aa4db34319a7b95e789065540347f895
Fixes: https://tracker.ceph.com/issues/44641
Signed-off-by: Nathan Cutler <ncutler@suse.com>