mon: Drop proposals that appear to come from our own rank
This can happen on startup when monitors are trying to guess
their ranks because not all IPs are known yet, and will get
sorted out as they finish probing each other.
Maybe it might also happen if a monitor is down and doesn't witness
removal of a peer from the monmap, but it should also get
updated as it probes so having the clashing rank ignore it
in elections shouldn't be a problem.

Signed-off-by: Greg Farnum <gfarnum@redhat.com>