mon: check if auth_mon if readable before auth_request
when performing mon thrash test, there is chance that a monitor is still
recovering when it tries to authorize a request and set the cap for the
connecting client, but if the key_server containes stale info and it's
not yet updated by the latest paxos being replayed, we might authorize the
client using the stale key ring and cap information.

in this change, check if the auth_server is readable before referencing
the key_server updated by it, return -EBUSY instead.

Fixes: https://tracker.ceph.com/issues/38082
Signed-off-by: Kefu Chai <kchai@redhat.com>