mgr/dashboard: Fix error when listing RBD while deleting or moving
This fix is based on a commit made to octopus, that fixed this problem in
octopus/master: 5627919c016785a27666b1992d336f8ff378a072.

Since the codebase is very different and a backport is not recommended or even
possible, I have created this commit with only the minimal code necessary.

Fixes: https://tracker.ceph.com/issues/42330

Signed-off-by: Tiago Melo <tmelo@suse.com>