mon/MgrMonitor: show different error message when disabling a dne module
this is a regression introduced by
0fb5d4956f60a249830a880170f8b28885143930

per Sage, we can enable a module when the mgr is not started yet using
"--force" command, but we should be allowed to disable it if the
non-existent module is added. this change just changes the condition and
error message so that the error message is printed if the module is not
loaded at all. because the operation should be idempotent, so the
operation is always finished without returning an error.

Fixes: https://tracker.ceph.com/issues/44728
Signed-off-by: Kefu Chai <kchai@redhat.com>