pybind/mgr: use six==1.14.0
this change is not cherry-picked from master. because in master, we
are using different ways to install six:

see
- 68527ab83a77d20b779759581ce0b8844991231b
- 1e07237d92da3bb5eba82b49f6276d17696e6985

because in teuthology we are using six.ensure_str, which was added in
six 1.12.0, see https://github.com/benjaminp/six/blob/1.12.0/CHANGES ,
we cannot continue using six 1.11.0, as a result, we need switch over to
six>1.12.0. since the latest stable version of six is now 1.14.0, let's
just use it.

Signed-off-by: Kefu Chai <kchai@redhat.com>