snap_schedule: add metadata fields, refactor Schedule, multiple Schedules per path
This commit adds a bunch of interesting metadata fields to the DB and
refactors the Schedule class so that it concentrates all the DB queries
in on place. Also enables to set multiple schedules per path as long as
the (repeat interval, start time) pairs are unique.

Signed-off-by: Jan Fajerski <jfajerski@suse.com>
(cherry picked from commit 92f21b1a2aed62416698bfc302fe816dce181ca1)