qa,test: decode return val of base64.b64encode()
base64.b64encode() returns bytes in Python3, and we will pass it as a
commandline parameter later on, which will be quoted. so the command
line can be consumed by shell. so before sending it to shell, we should
convert it to string.

Signed-off-by: Kefu Chai <kchai@redhat.com>
(cherry picked from commit 91b49ab39d7377e0816fa3c44e2cabed9ce14f12)

Conflicts:
	qa/tasks/s3tests_java.py: trivial resolution