qa/tasks: decode bytes returned by base64.b64encode()
we use the return value from `base64.b64encode()` as a string, this works
in python2, but in python3, instead of a string, `base64.b64encode()`
returns bytes. so for instance, if we try to compose command line
arguments using `pipes.quote()` by passing a byte instance, we will have
```
TypeError: cannot use a string pattern on a bytes-like object
```

in this change, the retval is always decoded using ASCII, because base64
only returns ASCII strings when encoding.

this change is not cherry-picked from master, as the changed files do
not exist in master anymore.

Signed-off-by: Kefu Chai <kchai@redhat.com>