mgr/cephadm: Add migration to the new scheduler.
New scheduler that takes PlacementSpec as the bound and not as recommendation.

Which means, we have to make sure, we're not removing any daemons directly after
upgrading to the new scheduler.

There is a potential race here:

1. user updates his spec to remove daemons
2. mgr gets upgrades to new scheduler, before the old scheduler removed the daemon
3. now, we're converting the spec to explicit placement, thus reverting (1.)

I think this is ok.

Signed-off-by: Sebastian Wagner <sebastian.wagner@suse.com>
(cherry picked from commit 074745217703ab57692fe8234a82130dc49179d6)