suite: allow override architecture
By default for generation job configuration and filtering
them out there is 'arch' field is used determined automatically
via request to paddles database for give machine_type.

This makes it possible to override the 'arch' value with
teuthology-suite '--arch' parameter. This is only useful
at the moment just when a user don't one to make any request
to paddles.
Originally the arch is used to filter out the suites
which are not supposed to be queued on the given nodes.

In future we probably need to have tests with heterogeneous
configuration which will use multiple architectures.

Signed-off-by: Kyr Shatskyy <kyrylo.shatskyy@suse.com>