mgr/dashboard: Improve Summary's subscribe methods
Now subscribe only emits if the value is not undefined,
removing the need to do the validation after each subscribe.

Add a new subscribeOnce method, which replaces getCurrentSummary.
Like the previous one, this will only emit when there is a non undefined value,
and will only emit once.
After it emits, the subscription is closed automatically.

Fixes: https://tracker.ceph.com/issues/36563

Signed-off-by: Tiago Melo <tmelo@suse.com>