qa/suites/fs/upgrade: disable mon_pg_warn_min_per_osd
if we upgrade from luminous, then a luminous monitor won't have the
1ac34a5ea3d1aca299b02e574b295dd4bf6167f4 or its backport, but after
upgrading to nautilus, mgr detects this issue and translates it to
a warning like

2020-05-21T05:33:26.934 DEBUG:teuthology.misc:Ceph health: HEALTH_WARN
too few PGs per OSD (12 < min 30)

if the number of pg per osd is too low. so we need to disable this
warning as we did in later releases and in master.

this change is not cherry-picked from master, as we don't do
upgrade test from luminous to master.

Signed-off-by: Kefu Chai <kchai@redhat.com>