cephadm: add --registry-{password,username} options
allow user to log into container registry with the specified
credentials. this helps user to address the issue if the registry is
protected by authentication machinary.

a global variable of `container_registry_credentials` is introduced. it
is an instance of `ContainerRegistryCredentials` which serves as an
decorator and also as a context manager.

Fixes: https://tracker.ceph.com/issues/44886
Signed-off-by: Kefu Chai <kchai@redhat.com>