RGW: Alias ES permissions to permissions.keyword
When `rgw_keystone_implicit_tenants = true` and using the Elasticsearch
sync module the search module when searching for `permissions` don't
work for buckets.

With `rgw_keystone_implicit_tenants` enabled buckets are stored as
`PROJECT_ID$BUCKET_NAME`. And when using permissions it splits by the
'$' and treats the bucket name as 2 word.

If we change permissions to permissions.keyword, then it'll be treated
as a single word.

This patch adds and alias for permissions to map to permissions.keyword.

Before the patch:

  $ radosgw-es 'permissions == projectname$bucketname'
  {"query":{"term":{"permissions":"projectname$bucketname"}}}

afterwards:

  $ radosgw-es 'permissions == projectname$bucketname'
  {"query":{"term":{"permissions.keyword":"projectname$bucketname"}}}

Fixes: https://tracker.ceph.com/issues/45607
Signed-off-by: Matthew Oliver <moliver@suse.com>